The First exercise of app architecture

+++ app/src/main/java/com/example/networkcall3/RepoData.kt
@@ -1,7 +1,7 @@
 package com.example.networkcall3
 
 data class RepoData(
-    val accessibility: Int,
+    val accessibility: Double,
     val activity: String,
     val key: String,
     val link: String,
